refactor: restructure timesheet import components and logic
- Introduced new components for file import, parsed file summary, and preview table to enhance modularity and readability.
- Consolidated existing logic into a custom hook for managing timesheet import state and operations.
- Improved user experience with editable cells for date and time inputs, ensuring proper formatting and validation.
- Added overlap confirmation and unresolved worker matching panels to handle edge cases during import.
- Removed deprecated functions and streamlined the import process for better maintainability.
