+++ app/dashboard/timesheet/import/_components/amber-callout.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+
+export function AmberCallout({
+    title,
+    children,
+}: {
+    title: string;
+    children: React.ReactNode;
+}) {
+    return (
+        <div className="rounded-md border border-amber-300 bg-amber-50 p-3 text-sm text-amber-900 dark:border-amber-900/60 dark:bg-amber-950/30 dark:text-amber-200">
+            <p className="font-medium">{title}</p>
+            {children}
+        </div>
+    );
+}
