+++ app/dashboard/timesheet/import/timesheet-import-client.tsx
@@ -1,12 +1,5 @@
 "use client";
 
-import * as React from "react";
-import * as XLSX from "@e965/xlsx";
-
-import { importAttendRecordTimesheet } from "./import-attend-record-timesheet";
-import type { OverlapEntry } from "@/services/timesheet/import-attend-record-timesheet";
-import { SelectSearch } from "@/components/ui/SelectSearch";
-import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
 import {
     Card,
@@ -16,435 +9,49 @@ import {
     CardTitle,
 } from "@/components/ui/card";
 import { FormPageLayout } from "@/components/form-page-layout";
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow,
-} from "@/components/ui/table";
-import { Plus, Trash2, TriangleAlert, Upload } from "lucide-react";
-import { parseAttendRecord } from "@/utils/payroll/parse-attendrecord";
-import { formatClockIntervalHm } from "@/utils/payroll/payroll-utils";
-import type { AttendRecordOutput } from "@/utils/payroll/parse-attendrecord";
-import {
-    editableRowsToAttendRecord,
-    flattenForPreview,
-    groupRowsByWorker,
-    hasRowError,
-    isPreviewDateInvalid,
-    isPreviewTimeInInvalid,
-    isPreviewTimeOutInvalid,
-    type TimesheetImportPreviewRow,
-} from "@/services/timesheet/import-preview";
-import { formatDmyInput } from "@/utils/time/calendar-date";
-import {
-    resolveTimesheetImportWorkerMatches,
-    type TimesheetImportWorker,
-} from "./worker-matching";
+import { hasRowError } from "@/services/timesheet/import-preview";
 
-const ACCEPTED_TYPES = [
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-    "application/vnd.ms-excel",
-    "text/csv",
-];
-
-const ACCEPTED_EXTENSIONS = [".xlsx", ".xls", ".csv"];
-
-function isValidFile(file: File): boolean {
-    return (
-        ACCEPTED_TYPES.includes(file.type) ||
-        ACCEPTED_EXTENSIONS.some((ext) => file.name.toLowerCase().endsWith(ext))
-    );
-}
-
-/** Enforce HH:MM format as user types - digits only, auto-insert colon */
-function formatTimeInput(raw: string): string {
-    const digits = raw.replace(/\D/g, "").slice(0, 4);
-    if (digits.length <= 2) return digits;
-    return `${digits.slice(0, 2)}:${digits.slice(2)}`;
-}
-
-function placeCaretAtEnd(el: HTMLElement) {
-    const range = document.createRange();
-    const sel = window.getSelection();
-    range.selectNodeContents(el);
-    range.collapse(false);
-    sel?.removeAllRanges();
-    sel?.addRange(range);
-}
-
-function AmberCallout({
-    title,
-    children,
-}: {
-    title: string;
-    children: React.ReactNode;
-}) {
-    return (
-        <div className="rounded-md border border-amber-300 bg-amber-50 p-3 text-sm text-amber-900 dark:border-amber-900/60 dark:bg-amber-950/30 dark:text-amber-200">
-            <p className="font-medium">{title}</p>
-            {children}
-        </div>
-    );
-}
-
-function ContentEditableDateCell(props: {
-    rowIndex: number;
-    field: "dateIn" | "dateOut";
-    value: string;
-    invalid: boolean;
-    onCommit: (
-        rowIndex: number,
-        field: "dateIn" | "dateOut",
-        value: string,
-    ) => void;
-}) {
-    const { rowIndex, field, value, invalid, onCommit } = props;
-    return (
-        <div
-            contentEditable
-            suppressContentEditableWarning
-            onInput={(e) => {
-                const el = e.currentTarget;
-                const formatted = formatDmyInput(el.textContent ?? "");
-                if (el.textContent !== formatted) {
-                    el.textContent = formatted;
-                    placeCaretAtEnd(el);
-                }
-            }}
-            onBlur={(e) =>
-                onCommit(rowIndex, field, e.currentTarget.textContent ?? "")
-            }
-            onPaste={(e) => {
-                e.preventDefault();
-                const text = e.clipboardData.getData("text/plain");
-                const formatted = formatDmyInput(text.replace(/\D/g, ""));
-                document.execCommand("insertText", false, formatted);
-            }}
-            className={`min-w-28 rounded px-2 py-1.5 outline-none focus:ring-2 focus:ring-ring focus:ring-inset ${invalid ? "text-destructive" : ""}`}>
-            {value}
-        </div>
-    );
-}
-
-function ContentEditableTimeCell(props: {
-    rowIndex: number;
-    field: "timeIn" | "timeOut";
-    displayText: string;
-    invalid: boolean;
-    onCommit: (
-        rowIndex: number,
-        field: "timeIn" | "timeOut",
-        value: string,
-    ) => void;
-    /** Time-out column: em dash placeholder and blur normalization. */
-    emptyPlaceholderEmDash?: boolean;
-}) {
-    const {
-        rowIndex,
-        field,
-        displayText,
-        invalid,
-        onCommit,
-        emptyPlaceholderEmDash = false,
-    } = props;
-
-    return (
-        <div
-            contentEditable
-            suppressContentEditableWarning
-            onInput={(e) => {
-                const el = e.currentTarget;
-                if (emptyPlaceholderEmDash) {
-                    const raw = (el.textContent ?? "").replace(/—/g, "");
-                    const formatted = raw === "" ? "" : formatTimeInput(raw);
-                    if (el.textContent !== formatted) {
-                        el.textContent = formatted || "—";
-                        if (formatted) placeCaretAtEnd(el);
-                    }
-                } else {
-                    const formatted = formatTimeInput(el.textContent ?? "");
-                    if (el.textContent !== formatted) {
-                        el.textContent = formatted;
-                        placeCaretAtEnd(el);
-                    }
-                }
-            }}
-            onBlur={(e) => {
-                const v = (e.currentTarget.textContent ?? "").trim();
-                if (emptyPlaceholderEmDash) {
-                    onCommit(rowIndex, field, v === "—" ? "" : v);
-                } else {
-                    onCommit(rowIndex, field, v);
-                }
-            }}
-            onPaste={(e) => {
-                e.preventDefault();
-                const text = e.clipboardData.getData("text/plain");
-                const formatted = formatTimeInput(text.replace(/\D/g, ""));
-                document.execCommand("insertText", false, formatted);
-            }}
-            className={`min-w-20 rounded px-2 py-1.5 outline-none focus:ring-2 focus:ring-ring focus:ring-inset ${invalid ? "text-destructive" : ""}`}>
-            {displayText}
-        </div>
-    );
-}
+import { ImportFileDropzone } from "./_components/import-file-dropzone";
+import { ImportParsedFileSummary } from "./_components/import-parsed-file-summary";
+import { ImportPreviewTable } from "./_components/import-preview-table";
+import { ImportSubmitResult } from "./_components/import-submit-result";
+import { OverlapConfirmationPanel } from "./_components/overlap-confirmation-panel";
+import { UnresolvedWorkerMatchingPanel } from "./_components/unresolved-worker-matching-panel";
+import type { TimesheetImportWorker } from "./worker-matching";
+import { useTimesheetImport } from "./use-timesheet-import";
 
 export function TimesheetImportClient({
     workers,
 }: {
     workers: TimesheetImportWorker[];
 }) {
-    const [isDragging, setIsDragging] = React.useState(false);
-    const [file, setFile] = React.useState<File | null>(null);
-    const [parsedData, setParsedData] =
-        React.useState<AttendRecordOutput | null>(null);
-    const [editableRows, setEditableRows] = React.useState<
-        TimesheetImportPreviewRow[]
-    >([]);
-    const [error, setError] = React.useState<string | null>(null);
-    const [submitResult, setSubmitResult] = React.useState<{
-        imported?: number;
-        skipped?: number;
-        errors?: string[];
-    } | null>(null);
-    const [overlapResult, setOverlapResult] = React.useState<
-        OverlapEntry[] | null
-    >(null);
-    const [pending, setPending] = React.useState(false);
-    const [manualWorkerMatches, setManualWorkerMatches] = React.useState<
-        Record<string, string>
-    >({});
-
-    const handleParse = React.useCallback(
-        async (file: File) => {
-            setError(null);
-            setSubmitResult(null);
-            setOverlapResult(null);
-            setManualWorkerMatches({});
-            try {
-                const data = await file.arrayBuffer();
-                const workbook = XLSX.read(data, {
-                    type: "array",
-                    cellDates: true,
-                });
-                const sheetName = workbook.SheetNames[0];
-                if (!sheetName) {
-                    setError("Workbook has no sheets.");
-                    setParsedData(null);
-                    setFile(null);
-                    return;
-                }
-                const ws = workbook.Sheets[sheetName];
-                const rows = XLSX.utils.sheet_to_json(ws, {
-                    header: 1,
-                    defval: null,
-                    raw: false,
-                }) as (string | number | null)[][];
-
-                const result = parseAttendRecord(rows);
-                if (!result.workers.length) {
-                    setError(
-                        "Unrecognized format — expected AttendRecord-style Excel",
-                    );
-                    setParsedData(null);
-                    setFile(null);
-                    return;
-                }
-                setParsedData(result);
-                setEditableRows(flattenForPreview(result, workers, {}));
-                setFile(file);
-            } catch (err) {
-                setError(
-                    err instanceof Error
-                        ? err.message
-                        : "Failed to parse Excel file",
-                );
-                setParsedData(null);
-                setFile(null);
-            }
-        },
-        [workers],
-    );
-
-    React.useEffect(() => {
-        if (!parsedData) return;
-        setEditableRows(
-            flattenForPreview(parsedData, workers, manualWorkerMatches),
-        );
-    }, [parsedData, workers, manualWorkerMatches]);
-
-    const updateEditableRow = React.useCallback(
-        (
-            rowIndex: number,
-            field: "workerName" | "dateIn" | "dateOut" | "timeIn" | "timeOut",
-            value: string,
-        ) => {
-            setEditableRows((prev) => {
-                const next = [...prev];
-                const row = { ...next[rowIndex]! };
-                row[field] = value;
-                next[rowIndex] = row;
-                return next;
-            });
-        },
-        [],
-    );
-
-    const addRowForWorker = React.useCallback((workerName: string) => {
-        setEditableRows((prev) => {
-            const firstIdx = prev.findIndex((r) => r.workerName === workerName);
-            if (firstIdx === -1) return prev;
-            const newRow: TimesheetImportPreviewRow = {
-                workerName,
-                dateIn: "",
-                dateOut: "",
-                timeIn: "",
-                timeOut: "",
-            };
-            const next = [...prev];
-            next.splice(firstIdx + 1, 0, newRow);
-            return next;
-        });
-    }, []);
-
-    const handleSubmit = React.useCallback(
-        async (mode?: "skip" | "force") => {
-            if (!parsedData || editableRows.length === 0) return;
-            setError(null);
-            if (mode == null) {
-                setSubmitResult(null);
-            }
-            setPending(true);
-            try {
-                const workerMatchResult = resolveTimesheetImportWorkerMatches({
-                    rows: editableRows,
-                    workers,
-                    manualMatchesByImportedName: manualWorkerMatches,
-                });
-                const resolvedWorkerNamesByImportedName = new Map(
-                    workerMatchResult.groups
-                        .filter((group) => group.resolvedWorker != null)
-                        .map((group) => [
-                            group.importedName,
-                            group.resolvedWorker!.name,
-                        ]),
-                );
-                const dataToImport = editableRowsToAttendRecord(
-                    editableRows,
-                    {
-                        attendanceDate: parsedData.attendanceDate,
-                        tablingDate: parsedData.tablingDate,
-                    },
-                    resolvedWorkerNamesByImportedName,
-                );
-                const result = await importAttendRecordTimesheet(
-                    dataToImport,
-                    mode,
-                );
-                if ("error" in result) {
-                    setError(result.error);
-                    return;
-                }
-                if (result.status === "confirmation_required") {
-                    setOverlapResult(result.overlaps);
-                    return;
-                }
-                setOverlapResult(null);
-                setSubmitResult({
-                    imported: result.imported,
-                    skipped: result.skipped,
-                    errors: result.errors,
-                });
-                if (result.imported > 0) {
-                    setFile(null);
-                    setParsedData(null);
-                    setEditableRows([]);
-                }
-            } catch (err) {
-                setError(err instanceof Error ? err.message : "Import failed");
-            } finally {
-                setPending(false);
-            }
-        },
-        [parsedData, editableRows, workers, manualWorkerMatches],
-    );
-
-    const onDrop = React.useCallback(
-        (e: React.DragEvent) => {
-            e.preventDefault();
-            setIsDragging(false);
-            const dropped = e.dataTransfer.files[0];
-            if (dropped && isValidFile(dropped)) {
-                handleParse(dropped);
-            } else {
-                setError("Please drop an Excel file (.xlsx, .xls) or CSV");
-            }
-        },
-        [handleParse],
-    );
-
-    const onDragOver = React.useCallback((e: React.DragEvent) => {
-        e.preventDefault();
-        e.stopPropagation();
-        setIsDragging(true);
-    }, []);
-
-    const onDragLeave = React.useCallback((e: React.DragEvent) => {
-        e.preventDefault();
-        e.stopPropagation();
-        setIsDragging(false);
-    }, []);
-
-    const onFileInputChange = React.useCallback(
-        (e: React.ChangeEvent<HTMLInputElement>) => {
-            const selected = e.target.files?.[0];
-            if (selected && isValidFile(selected)) {
-                handleParse(selected);
-            }
-            e.target.value = "";
-        },
-        [handleParse],
-    );
-
-    const reset = React.useCallback(() => {
-        setFile(null);
-        setParsedData(null);
-        setEditableRows([]);
-        setError(null);
-        setSubmitResult(null);
-        setOverlapResult(null);
-        setManualWorkerMatches({});
-    }, []);
+    const {
+        isDragging,
+        file,
+        parsedData,
+        editableRows,
+        setEditableRows,
+        error,
+        submitResult,
+        overlapResult,
+        pending,
+        onDrop,
+        onDragOver,
+        onDragLeave,
+        onFileInputChange,
+        reset,
+        updateEditableRow,
+        addRowForWorker,
+        handleSubmit,
+        workerMatchResult,
+        workerMatchGroupsByImportedName,
+        activeWorkers,
+        setManualWorkerMatch,
+    } = useTimesheetImport(workers);
 
     const totalEntries = editableRows.length;
-    const workerMatchResult = React.useMemo(
-        () =>
-            resolveTimesheetImportWorkerMatches({
-                rows: editableRows,
-                workers,
-                manualMatchesByImportedName: manualWorkerMatches,
-            }),
-        [editableRows, workers, manualWorkerMatches],
-    );
-    const workerMatchGroupsByImportedName = React.useMemo(
-        () =>
-            new Map(
-                workerMatchResult.groups.map((group) => [
-                    group.importedName,
-                    group,
-                ]),
-            ),
-        [workerMatchResult.groups],
-    );
     const hasUnresolvedWorkerMatches =
         workerMatchResult.unresolvedNames.length > 0;
-    const activeWorkers = React.useMemo(
-        () => workers.filter((worker) => worker.status === "Active"),
-        [workers],
-    );
 
     return (
         <FormPageLayout
@@ -459,32 +66,13 @@ export function TimesheetImportClient({
                     </CardDescription>
                 </CardHeader>
                 <CardContent>
-                    <label
+                    <ImportFileDropzone
+                        isDragging={isDragging}
                         onDrop={onDrop}
                         onDragOver={onDragOver}
                         onDragLeave={onDragLeave}
-                        className={`
-                            flex min-h-[200px] cursor-pointer flex-col items-center justify-center
-                            gap-3 rounded-lg border-2 border-dashed transition-colors
-                            hover:border-primary/50 hover:bg-muted/50
-                            ${isDragging ? "border-primary bg-muted/50" : "border-muted-foreground/25"}
-                        `}>
-                        <input
-                            type="file"
-                            accept=".xlsx,.xls,application/vnd.openxmlformats-officedocument.spreadsheetml.sheet,application/vnd.ms-excel"
-                            onChange={onFileInputChange}
-                            className="sr-only"
-                        />
-                        <Upload className="text-muted-foreground size-10" />
-                        <span className="text-muted-foreground text-sm">
-                            {isDragging
-                                ? "Drop file here"
-                                : "Drag and drop or click to upload"}
-                        </span>
-                        <span className="text-muted-foreground text-xs">
-                            .xlsx, .xls
-                        </span>
-                    </label>
+                        onFileInputChange={onFileInputChange}
+                    />
 
                     {error && (
                         <p className="text-destructive mt-3 text-sm">{error}</p>
@@ -492,394 +80,45 @@ export function TimesheetImportClient({
 
                     {file && parsedData && (
                         <div className="mt-4 space-y-4">
-                            <div className="flex flex-wrap items-center gap-4">
-                                <p className="font-medium">{file.name}</p>
-                                <button
-                                    type="button"
-                                    onClick={reset}
-                                    className="text-primary hover:underline text-sm">
-                                    Upload a different file
-                                </button>
-                            </div>
-                            <div className="text-muted-foreground text-sm">
-                                <p>
-                                    Attendance period:{" "}
-                                    {parsedData.attendanceDate.startDate} ~{" "}
-                                    {parsedData.attendanceDate.endDate}
-                                </p>
-                                {parsedData.tablingDate && (
-                                    <p>
-                                        Tabling date: {parsedData.tablingDate}
-                                    </p>
-                                )}
-                                <p>
-                                    {parsedData.workers.length} worker
-                                    {parsedData.workers.length !== 1 ? "s" : ""}
-                                    , {totalEntries} entr
-                                    {totalEntries !== 1 ? "ies" : "y"} total
-                                </p>
-                            </div>
-                            {hasUnresolvedWorkerMatches && (
-                                <AmberCallout title="Unresolved worker matches">
-                                    <p className="mt-1">
-                                        {
-                                            workerMatchResult.unresolvedNames
-                                                .length
-                                        }{" "}
-                                        worker{" "}
-                                        {workerMatchResult.unresolvedNames
-                                            .length !== 1
-                                            ? "names"
-                                            : "name"}{" "}
-                                        must match an active Worker before
-                                        upload.
-                                    </p>
-                                    <ul className="mt-2 list-inside list-disc">
-                                        {workerMatchResult.unresolvedNames.map(
-                                            (workerName) => (
-                                                <li key={workerName}>
-                                                    {workerName}
-                                                </li>
-                                            ),
-                                        )}
-                                    </ul>
-                                </AmberCallout>
+                            <ImportParsedFileSummary
+                                fileName={file.name}
+                                parsedData={parsedData}
+                                totalEntries={totalEntries}
+                                onReset={reset}
+                            />
+                            <UnresolvedWorkerMatchingPanel
+                                unresolvedNames={
+                                    workerMatchResult.unresolvedNames
+                                }
+                            />
+                            {overlapResult != null && (
+                                <OverlapConfirmationPanel
+                                    overlaps={overlapResult}
+                                    pending={pending}
+                                    onSkipDuplicates={() => {
+                                        void handleSubmit("skip");
+                                    }}
+                                    onImportAll={() => {
+                                        void handleSubmit("force");
+                                    }}
+                                />
                             )}
-                            {overlapResult != null &&
-                                overlapResult.length > 0 && (
-                                    <AmberCallout title="Overlapping timesheet dates">
-                                        <p className="mt-1">
-                                            {overlapResult.reduce(
-                                                (sum, o) =>
-                                                    sum + o.existingCount,
-                                                0,
-                                            )}{" "}
-                                            entr
-                                            {overlapResult.reduce(
-                                                (sum, o) =>
-                                                    sum + o.existingCount,
-                                                0,
-                                            ) !== 1
-                                                ? "ies"
-                                                : "y"}{" "}
-                                            already exist for{" "}
-                                            {
-                                                new Set(
-                                                    overlapResult.map(
-                                                        (o) => o.workerName,
-                                                    ),
-                                                ).size
-                                            }{" "}
-                                            worker
-                                            {new Set(
-                                                overlapResult.map(
-                                                    (o) => o.workerName,
-                                                ),
-                                            ).size !== 1
-                                                ? "s"
-                                                : ""}{" "}
-                                            on the selected dates. Choose how to
-                                            proceed.
-                                        </p>
-                                        <ul className="mt-2 max-h-32 list-inside list-disc overflow-y-auto">
-                                            {overlapResult.map((o, idx) => (
-                                                <li
-                                                    key={`${o.workerName}-${o.dateIn}-${idx}`}>
-                                                    {o.workerName} — {o.dateIn}{" "}
-                                                    ({o.existingCount} existing)
-                                                </li>
-                                            ))}
-                                        </ul>
-                                        <div className="mt-3 flex flex-wrap gap-2">
-                                            <Button
-                                                type="button"
-                                                variant="outline"
-                                                size="sm"
-                                                disabled={pending}
-                                                onClick={() => {
-                                                    void handleSubmit("skip");
-                                                }}>
-                                                Skip duplicates
-                                            </Button>
-                                            <Button
-                                                type="button"
-                                                size="sm"
-                                                disabled={pending}
-                                                onClick={() => {
-                                                    void handleSubmit("force");
-                                                }}>
-                                                Import all anyway
-                                            </Button>
-                                        </div>
-                                    </AmberCallout>
-                                )}
-                            <div className="rounded-md border overflow-x-auto max-h-[70vh] overflow-y-auto">
-                                <Table>
-                                    <TableHeader>
-                                        <TableRow>
-                                            <TableHead>Worker</TableHead>
-                                            <TableHead>Date in</TableHead>
-                                            <TableHead>Time in</TableHead>
-                                            <TableHead>Date out</TableHead>
-                                            <TableHead>Time out</TableHead>
-                                            <TableHead>Hours</TableHead>
-                                            <TableHead className="w-10" />
-                                        </TableRow>
-                                    </TableHeader>
-                                    <TableBody>
-                                        {groupRowsByWorker(editableRows).map(
-                                            (
-                                                {
-                                                    row,
-                                                    flatIndex: i,
-                                                    isFirstInGroup,
-                                                },
-                                                idx,
-                                            ) => (
-                                                <TableRow
-                                                    key={i}
-                                                    className={
-                                                        [
-                                                            hasRowError(row) &&
-                                                                "bg-destructive/10 hover:bg-destructive/25",
-                                                            isFirstInGroup &&
-                                                                idx > 0 &&
-                                                                "border-t-2 border-muted-foreground/20",
-                                                        ]
-                                                            .filter(Boolean)
-                                                            .join(" ") ||
-                                                        undefined
-                                                    }>
-                                                    <TableCell
-                                                        className={cn(
-                                                            "p-1 min-w-[180px]",
-                                                            !isFirstInGroup &&
-                                                                "pl-8",
-                                                        )}>
-                                                        {isFirstInGroup ? (
-                                                            <div className="flex items-center gap-2">
-                                                                {workerMatchGroupsByImportedName.get(
-                                                                    row.workerName,
-                                                                )
-                                                                    ?.resolvedWorker ==
-                                                                null ? (
-                                                                    <span
-                                                                        role="img"
-                                                                        aria-label="Imported name does not match an active worker — select a worker to continue"
-                                                                        title="Imported name does not match an active worker — select a worker to continue"
-                                                                        className="ml-2 flex shrink-0 items-center justify-center">
-                                                                        <TriangleAlert className="text-destructive size-6" />
-                                                                    </span>
-                                                                ) : null}
-                                                                <div className="flex-1 space-y-1">
-                                                                    <SelectSearch
-                                                                        options={activeWorkers.map(
-                                                                            (
-                                                                                w,
-                                                                            ) => ({
-                                                                                value: w.id,
-                                                                                label: w.name,
-                                                                            }),
-                                                                        )}
-                                                                        value={
-                                                                            workerMatchGroupsByImportedName.get(
-                                                                                row.workerName,
-                                                                            )
-                                                                                ?.resolvedWorker
-                                                                                ?.id ??
-                                                                            ""
-                                                                        }
-                                                                        onChange={(
-                                                                            id,
-                                                                        ) =>
-                                                                            setManualWorkerMatches(
-                                                                                (
-                                                                                    prev,
-                                                                                ) => ({
-                                                                                    ...prev,
-                                                                                    [row.workerName]:
-                                                                                        id,
-                                                                                }),
-                                                                            )
-                                                                        }
-                                                                        name={`worker-${i}`}
-                                                                        placeholder={
-                                                                            row.workerName ||
-                                                                            "Select worker"
-                                                                        }
-                                                                        searchPlaceholder="Search workers…"
-                                                                        emptyText="No workers found."
-                                                                    />
-                                                                    <p className="text-muted-foreground text-xs">
-                                                                        Source:{" "}
-                                                                        {
-                                                                            row.workerName
-                                                                        }
-                                                                    </p>
-                                                                </div>
-                                                                <button
-                                                                    type="button"
-                                                                    onClick={() =>
-                                                                        addRowForWorker(
-                                                                            row.workerName,
-                                                                        )
-                                                                    }
-                                                                    className="text-muted-foreground hover:text-primary shrink-0 rounded p-1"
-                                                                    aria-label={`Add row for ${row.workerName}`}>
-                                                                    <Plus className="size-4" />
-                                                                </button>
-                                                            </div>
-                                                        ) : (
-                                                            <span className="text-muted-foreground/50">
-                                                                └
-                                                            </span>
-                                                        )}
-                                                    </TableCell>
-                                                    <TableCell className="p-1">
-                                                        <ContentEditableDateCell
-                                                            rowIndex={i}
-                                                            field="dateIn"
-                                                            value={row.dateIn}
-                                                            invalid={isPreviewDateInvalid(
-                                                                row.dateIn,
-                                                            )}
-                                                            onCommit={
-                                                                updateEditableRow
-                                                            }
-                                                        />
-                                                    </TableCell>
-                                                    <TableCell className="p-1">
-                                                        <ContentEditableTimeCell
-                                                            rowIndex={i}
-                                                            field="timeIn"
-                                                            displayText={
-                                                                row.timeIn
-                                                            }
-                                                            invalid={isPreviewTimeInInvalid(
-                                                                row.timeIn,
-                                                            )}
-                                                            onCommit={
-                                                                updateEditableRow
-                                                            }
-                                                        />
-                                                    </TableCell>
-                                                    <TableCell className="p-1">
-                                                        <ContentEditableDateCell
-                                                            rowIndex={i}
-                                                            field="dateOut"
-                                                            value={row.dateOut}
-                                                            invalid={isPreviewDateInvalid(
-                                                                row.dateOut,
-                                                            )}
-                                                            onCommit={
-                                                                updateEditableRow
-                                                            }
-                                                        />
-                                                    </TableCell>
-                                                    <TableCell className="p-1">
-                                                        <ContentEditableTimeCell
-                                                            rowIndex={i}
-                                                            field="timeOut"
-                                                            displayText={
-                                                                row.timeOut ||
-                                                                "—"
-                                                            }
-                                                            invalid={isPreviewTimeOutInvalid(
-                                                                row.timeOut,
-                                                            )}
-                                                            onCommit={
-                                                                updateEditableRow
-                                                            }
-                                                            emptyPlaceholderEmDash
-                                                        />
-                                                    </TableCell>
-                                                    {(() => {
-                                                        const hours =
-                                                            formatClockIntervalHm(
-                                                                row.dateIn,
-                                                                row.timeIn,
-                                                                row.dateOut,
-                                                                row.timeOut,
-                                                            );
-                                                        return (
-                                                            <TableCell
-                                                                className={`p-1 tabular-nums ${hours === "0h" ? "text-destructive" : "text-muted-foreground"}`}>
-                                                                {hours}
-                                                            </TableCell>
-                                                        );
-                                                    })()}
-                                                    <TableCell className="w-10 p-1">
-                                                        <button
-                                                            type="button"
-                                                            onClick={() =>
-                                                                setEditableRows(
-                                                                    (prev) =>
-                                                                        prev.filter(
-                                                                            (
-                                                                                _,
-                                                                                idx,
-                                                                            ) =>
-                                                                                idx !==
-                                                                                i,
-                                                                        ),
-                                                                )
-                                                            }
-                                                            className="text-muted-foreground hover:text-destructive rounded p-1"
-                                                            aria-label="Delete entry">
-                                                            <Trash2 className="size-4" />
-                                                        </button>
-                                                    </TableCell>
-                                                </TableRow>
-                                            ),
-                                        )}
-                                    </TableBody>
-                                </Table>
-                            </div>
+                            <ImportPreviewTable
+                                editableRows={editableRows}
+                                onEditableRowsChange={setEditableRows}
+                                workerMatchGroupsByImportedName={
+                                    workerMatchGroupsByImportedName
+                                }
+                                activeWorkers={activeWorkers}
+                                onManualWorkerMatch={setManualWorkerMatch}
+                                onAddRowForWorker={addRowForWorker}
+                                onUpdateRow={updateEditableRow}
+                            />
                         </div>
                     )}
+
                     {submitResult && (
-                        <div className="mt-4 rounded-lg border p-4">
-                            {submitResult.imported != null &&
-                                submitResult.imported > 0 && (
-                                    <p className="text-emerald-600 dark:text-emerald-400">
-                                        Imported {submitResult.imported}{" "}
-                                        entries.
-                                        {submitResult.skipped != null &&
-                                            submitResult.skipped > 0 && (
-                                                <>
-                                                    {" "}
-                                                    Skipped{" "}
-                                                    {submitResult.skipped}{" "}
-                                                    duplicates.
-                                                </>
-                                            )}
-                                    </p>
-                                )}
-                            {submitResult.imported === 0 &&
-                                (submitResult.skipped ?? 0) > 0 && (
-                                    <p className="text-amber-600 dark:text-amber-400">
-                                        No new entries imported. Skipped{" "}
-                                        {submitResult.skipped} duplicates.
-                                    </p>
-                                )}
-                            {submitResult.errors &&
-                                submitResult.errors.length > 0 && (
-                                    <ul className="mt-2 list-inside list-disc text-sm text-amber-600 dark:text-amber-400">
-                                        {submitResult.errors
-                                            .slice(0, 5)
-                                            .map((e, i) => (
-                                                <li key={i}>{e}</li>
-                                            ))}
-                                        {submitResult.errors.length > 5 && (
-                                            <li>
-                                                …and{" "}
-                                                {submitResult.errors.length - 5}{" "}
-                                                more
-                                            </li>
-                                        )}
-                                    </ul>
-                                )}
-                        </div>
+                        <ImportSubmitResult submitResult={submitResult} />
                     )}
 
                     <div className="mt-4 flex justify-end">
